+++ google/CHANGELOG.md
@@ -2,9 +2,14 @@
 
 This file documents changes to the `workbench-google` library, including notes on how to upgrade to new versions.
 
+## 0.19
+- Moved `org.broadinstitute.dsde.workbench.google.GoogleKmsService` to `org.broadinstitute.dsde.workbench.google2.GoogleKmsService`
+
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google" % "0.19-TRAVIS-REPLACE-ME"`
+
 ## 0.18
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.18-d4e0782"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google" % "0.18-d4e0782"`
 
 ### Changed
 - Moved `SamModel`